Tests/LibWeb/Text/expected/hit_testing/visibility.txt

<DIV>
<BODY>
---
<DIV>
<BODY>
---
<BODY>
<HTML>
---
